+++ CHANGELOG.md
@@ -1,11 +1,16 @@
+# 0.9.64 Release:
+## Bug Fixes:
+- Spirit Guardians
+- Compelled Duel
+## Update Notes:
+- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet.
+  
 # 0.9.63 Release:
 ## New Content:
 - Twilight Domain Cleric: Twilight Shroud
 - Feat: Dual Wielder
 ## Bug Fixes:
 - Channel Divinity: Twilight Sanctuary
-## Update Notes:
-- Support for D&D system version 3.0.0 will be in a future update. Do not update your system yet.
   
 # 0.9.62 Release:
 ## New Content: